Combine the extract/apply deltas to one file so they're next to each other in the menu
